+++ .gitignore
@@ -1,3 +1,4 @@
 data/
 src/__pycache__/*
 test/__pycache__/*
+final-exam/__pycache__/*
